+++ code/src/components/Task.js
@@ -1,6 +1,7 @@
-import React, { Component } from 'react'
+import React from 'react'
 import { useDispatch } from 'react-redux'
 import { tasks } from 'reducers/tasks'
+import moment from 'moment'
 
 export const Task = ({ task }) => {
   const dispatch = useDispatch()
@@ -9,6 +10,8 @@ export const Task = ({ task }) => {
     <div>
       {console.log(task.text + task.id)}
       {task.id} {task.text} {task.display ? "Display" : "Ej display"} {task.complete ? "Klar" : "Ej klar"}
+      {moment(task.date).fromNow()}
+
       <button type="button" onClick={() => {
         dispatch(
           tasks.actions.removeTask(task.id))
@@ -21,10 +24,3 @@ export const Task = ({ task }) => {
   )
 }
 
-const mapStateToProps = (state) => ({
-
-})
-
-const mapDispatchToProps = {
-
-}
